+++ tests/e2e/utils/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## Added
 
+- `shopper.gotoMyAccount()` go to the /my-account/ page
 - `clickFilter()` util helper method that clicks on a list page filter
 - `moveAllItemsToTrash()` util helper method that checks every item in a list page and moves them to the trash
 - `createSimpleOrder( status )` component which accepts an order status string and creates a basic order with that status
@@ -14,6 +15,7 @@
 
 ## Changes
 
+- Deprecated `StoreOwnerFlow`, `CustomerFlow` in favour of `merchant`,`shopper`
 - `createSimpleOrder( status )` returns the ID of the order that was created
 
 # 0.1.1